+++ src/leotools/__init__.py
@@ -8,9 +8,7 @@ leotools
 - preproc: Satellite image preprocessing tools.
 """
 
-__version__ = "0.7.2"
-# with open("VERSION.txt") as f:
-#     __version__ = f.read()
+__version__ = "0.8.0"
 
 __author__ = "Rómeó Komolafe"