+++ src/leotools/constants.py
@@ -1,9 +1,6 @@
 """This module contains often used constants."""
 
-from pathlib import Path
-
 from rasterio.profiles import Profile
-from rasterio.crs import CRS
 
 ### Constants
 EOV = 'EPSG:23700'