+++ src/leotools/gistools.py
@@ -6,12 +6,14 @@ from math import ceil
 import numpy as np
 from osgeo import gdal
 import rasterio
-import rasterio.mask
+from rasterio.mask import mask
+from rasterio.merge import merge
 from rasterio.enums import Resampling
+from rasterio.warp import aligned_target, reproject
 import geopandas as gpd
-from shapely.geometry import Polygon
+from shapely.geometry import Polygon, box
 
-from leotools.basetools import load_files, ProcessTimer
+from .basetools import check_path, load_files, ProcessTimer
 
 ### Metadata manipulation
 
@@ -24,6 +26,11 @@ def get_tags(image):
     """Returns the GeoTiff tags of an image."""
     with rasterio.open(image) as src:
         return src.tags()
+
+def get_desc(image):
+    """Returns the decriptions of the image bands."""
+    with rasterio.open(image) as src:
+        return src.descriptions
 
 def strip_tags(image):
     """Removes the GeoTIFF tags from an image."""
@@ -109,7 +116,7 @@ def image_to_array(image, bands=None):
     with rasterio.open(image) as src:
         return src.read(bands)
 
-def array_to_image(array, output, sample_image=None, profile=None):
+def array_to_image(array, output, sample_image=None, profile=None, mode=0):
     """Exports a numpy array to an image with the properties of a sample file.
 
     The band count of the array is automatically detected. Dimensions should be
@@ -120,6 +127,8 @@ def array_to_image(array, output, sample_image=None, profile=None):
         output (path): Path to the output image.
         sample_image (path, optional): Path to the sample image.
         profile (dict, optional): A dict of additional metadata parameters.
+        mode (int, optional): Mode used for checking the output paths. Based on
+            the `check_path` function.
 
     Returns:
         None
@@ -127,6 +136,7 @@ def array_to_image(array, output, sample_image=None, profile=None):
 
     ### https://rasterio.readthedocs.io/en/latest/topics/image_processing.html#imageorder
     
+    check_path(Path(output).parent, mode)
     shape = array.shape
     len_shape = len(shape)
 
@@ -219,6 +229,35 @@ def mask_array(array, mask, nodata=0, invert=False):
 
     return masked_array
 
+def stack_arrays(arrays):
+    """Stacks multiple 2D or 3D arrays into a single 3D array.
+    
+    Args:
+        arrays (list): The arrays to stack.
+
+    Returns:
+        stacked_array (array): The stacked array.
+    """
+
+    # arrays = [i for i in arrays if len(i.shape) == 2 else np.expand_dims(i, 0)]
+    new_arrays = []
+    for i in arrays:
+        len_shape = len(i.shape)
+
+        ### Multi-band image shape = count, rows(y), cols(x)
+        if len_shape == 3:
+            new_arrays += [i]
+
+        ### Single-band image shape = rows(y), cols(x)
+        elif len_shape == 2:
+            new_arrays += [np.expand_dims(i, 0)]
+        
+        ### Last resort
+        else:
+            raise ValueError("The arrays sould be 2 or 3 dimensional.")
+
+    return np.concatenate(new_arrays)
+
 ### Band operations and NDXI calculation
 
 class BandSelector:
@@ -231,13 +270,15 @@ class BandSelector:
         self.src = src
 
     def __getitem__(self, index):
-        return self.src.read(index).astype('float32') ### Indexing from 1 #! dtype megadva
+        return self.src.read(index) ### Indexing from 1
+
+### Raster operations
 
 def ndiff(e1, e2):
     """Normalized difference function. Returns x, where x = (e1-e2) / (e1+e2)."""
     return (e1-e2) / (e1+e2)
 
-def extract_bands(image, output_dir, bands=None, dtype=None):
+def extract_bands(image, output_dir, bands=None, dtype=None, mode=0):
     """Extract bands or band combinations of an image.
 
     Args:
@@ -247,11 +288,14 @@ def extract_bands(image, output_dir, bands=None, dtype=None):
             output.
         dtype (str, optional): The datatype of the output images. Defaults to
             the datatype of the input.
+        mode (int, optional): Mode used for checking the output paths. Based on
+            the `check_path` function.
 
     Returns:
         None
     """
 
+    check_path(output_dir, mode)
     image = Path(image)
     output_dir = Path(output_dir)
 
@@ -292,84 +336,136 @@ def extract_bands(image, output_dir, bands=None, dtype=None):
         else:
             raise TypeError("Bands should be list, dict or None.")
 
-def stack_arrays(arrays):
-    """Stacks multiple 2D or 3D arrays into a single image.
-    
-    Args:
-        arrays (list): The arrays to stack.
-
-    Returns:
-        stacked_array (array): The stacked array.
-    """
-
-    # arrays = [i for i in arrays if len(i.shape) == 2 else np.expand_dims(i, 0)]
-    new_arrays = []
-    for i in arrays:
-        len_shape = len(i.shape)
-
-        ### Multi-band image shape = count, rows(y), cols(x)
-        if len_shape == 3:
-            new_arrays += [i]
-
-        ### Single-band image shape = rows(y), cols(x)
-        elif len_shape == 2:
-            new_arrays += [np.expand_dims(i, 0)]
-        
-        ### Last resort
-        else:
-            raise ValueError("The arrays sould be 2 or 3 dimensional.")
-
-    return np.concatenate(new_arrays)
-
-def stack_images(input_paths, output, band=0, bands_csv=None, recursive=False):
+def stack_images(input_paths, output, band=0, bands_csv=None, recursive=False, mode=0, verbose=False):
     """Stacks multiple images into a single image.
 
     Args:
         input_paths (path or list): A list of the source images.
-        output (path): The output image.
+        output (path): Output image.
         band (int, optional): Which band of the source images to use. Uses all
             bands of each image if 0.
         bands_csv (path, optional): A .csv file listing the source (image and
             band) for each band of the output image.
         recursive (bool, optional): Whether to scan the input directory
             recursively.
+        mode (int, optional): Mode used for checking the output paths. Based on
+            the `check_path` function.
+        verbose (bool, optional): Whether to print current process status.
 
     Returns:
         None
     """
 
+    check_path(Path(output).parent, mode)
     image_list = load_files(input_paths, '*.tif', recursive)
     dst_count = sum([get_profile(i)['count'] for i in image_list]) if not band else len(image_list)
     profile = get_profile(image_list[0])
     profile.update(count=dst_count)
-    at_band = 1
+    first_band = 1
+
+    if verbose:
+        print(f"Bands wirtten: 0/{dst_count}", end='\r')
     
     with rasterio.open(output, 'w', **profile) as dst:
         for i in image_list:
             with rasterio.open(i) as src:
                 src_count = 1 if band else src.count
-                dst_bands = at_band if band else list(range(at_band, at_band+src_count))
+                dst_bands = first_band if band else list(range(first_band, first_band+src_count))
                 dst.write(src.read(band or None), dst_bands)
 
                 ### Creating csv
                 if bands_csv:
                     with open(bands_csv, 'a') as f:
-                        lines = [f"{at_band+j},{i},{band or j+1}\n" for j in range(src_count)]
+                        lines = [f"{first_band+j},{i},{band or j+1}\n" for j in range(src_count)]
                         f.writelines(lines)
 
-                at_band += src_count
+                first_band += src_count
+
+            if verbose:
+                print(f"Bands wirtten: {first_band-1}/{dst_count}", end='\r')
+
+        if verbose:
+            print()
+
+def merge_images(input_paths, output, profile=None, round=0, recursive=False, mode=0, verbose=False):
+    """Merges multiple images into one.
+
+    Args:
+        input_paths (path or list): Images to work on.
+        output (path): Output image.
+        profile (dict, optional): A dict of additional image parameters.
+        round (number, optional): The grid to round the extent to. 
+        recursive (bool, optional): Whether to scan the input paths recursively.
+        mode (int, optional): Mode used for checking the output paths. Based on
+            the `check_path` function.
+        verbose (bool, optional): Whether to print current process status.
+
+    Returns:
+        None
+    """
+
+    check_path(Path(output).parent, mode)
+    image_list = load_files(input_paths, '*.tif', recursive)
+    opened_images = [rasterio.open(i) for i in image_list]
+
+    all_bounds = list(zip(*[i.bounds for i in opened_images]))
+    final_bounds = box(min(all_bounds[0]), min(all_bounds[1]), max(all_bounds[2]), max(all_bounds[3]))
+
+    dst_transform, dst_width, dst_height = calc_transform(
+        final_bounds,
+        opened_images[0].res[0],
+        opened_images[0].crs,
+        opened_images[0].crs,
+        round,
+    )
+
+    if not profile:
+        profile = {}
+
+    dst_profile = opened_images[0].profile
+
+    dst_profile.update(
+        **profile,
+        transform=dst_transform,
+        width=dst_width,
+        height=dst_height,
+    )
+
+    count = opened_images[0].count
+    with rasterio.open(output, 'w', **dst_profile) as dst:
+
+        ### Merging images
+        if verbose:
+            print(f"Bands wirtten: 0/{count}", end='\r')
+
+        for i in range(count):
+            ### https://rasterio.readthedocs.io/en/latest/api/rasterio.merge.html
+            dst_array, dst_transform = merge(opened_images, indexes=[i+1])
+            dst_array = dst_array[0, :, :]
+            dst.write(dst_array, i+1)
+            del dst_array
+
+            if verbose:
+                print(f"Bands wirtten: {i+1}/{count}", end='\r')
+        
+        if verbose:
+            print()
+
+        ### Closing opened images
+        for i in opened_images:
+            i.close()
 
 ### Making external files
 
-def make_ovr(image, levels=None, compress='zstd', compress_level=None, predictor=2):
+def make_ovr(image, levels=(2, 8, 32, 128), compress='zstd', compress_level=None, predictor=2):
     """Creates pyramid layers for an image in an external .ovr file.
     
     Args:
         image (path): The image to make an overview for.
         levels (tuple, optional): Overview levels.
         compress (str, optonal): Compression method used.
-        compress_level (int, optional): Compression level used. Uses default if
-            not provided.
+        compress_level (int, optional): Compression level used. Uses default for
+            the given compression if not provided.
         predictor (int): Predictor used.
 
     Returns:
@@ -385,7 +481,7 @@ def make_ovr(image, levels=None, compress='zstd', compress_level=None, predictor
         GDAL_NUM_THREADS=4,
     ):
         with rasterio.open(image, 'r+') as src:
-            src.build_overviews(levels or (2, 8, 32, 128), Resampling.average)
+            src.build_overviews(levels, Resampling.average)
 
 def make_aux(image):
     """Calculates image statistics and puts them into an .aux.xml file."""
@@ -422,7 +518,7 @@ def load_polygons(polygons):
     else:
         raise TypeError("Polygons should be paths or geometries.")
 
-def multi_mask(image, output, nodata=0, area_masks=None, bound_masks=None, all_touched=True, invert=False):
+def multi_mask(image, output, nodata=0, area_masks=None, bound_masks=None, all_touched=True, invert=False, mode=0):
     """Pixels of the raster that touch the mask shapes are filled with nodata.
 
     A mask can be:
@@ -440,10 +536,14 @@ def multi_mask(image, output, nodata=0, area_masks=None, bound_masks=None, all_t
             the geometries is masked out.
         invert (bool, optional): If True, pixels under the geometries are
             discarded, if False (default), they are kept.
+        mode (int, optional): Mode used for checking the output paths. Based on
+            the `check_path` function.
 
     Returns:
         None
     """
+
+    check_path(Path(output).parent, mode)
 
     ### Combining vector mask
     sum_mask = []
@@ -462,38 +562,162 @@ def multi_mask(image, output, nodata=0, area_masks=None, bound_masks=None, all_t
     ### https://rasterio.readthedocs.io/en/latest/api/rasterio.mask.html
     with rasterio.open(image) as src:
         profile = src.profile
-        out_img, transform = rasterio.mask.mask(src, sum_mask, all_touched=all_touched, invert=invert, nodata=nodata)
+        out_img, transform = mask(src, sum_mask, all_touched=all_touched, invert=invert, nodata=nodata)
 
     profile.update(transform=transform, nodata=nodata)
 
     with rasterio.open(output, 'w', **profile) as dst:
         dst.write(out_img)
 
-### Extent operations
+### Coordinate operations
 
-def round_to(num, resolution):
-    """Rounds number. A positive/negative resolution rounds up/down respectively."""
+def round_to(x, y):
+    """Rounds x to the closest multiple of y.
+    
+    A positive/negative y rounds up/down respectively.
+    """
 
-    return (ceil(num / resolution)) * resolution
+    return (ceil(x / y)) * y
 
-def round_extent(bounds, resolution):
-    """Rounds extent bounds to the resolution.
+def round_extent(bounds, grid):
+    """Rounds extent bounds to the grid.
     
     Args:
-        bounds (list): A list of extent bounds in this order:
-            xmin, ymin, xmax, ymax.
-        resolution (int): The resolution to round to. A positive/negative
-            resolution expands/shrinks the extent respectively.
+        bounds (list): Extent bounds in this order: xmin, ymin, xmax, ymax.
+        grid (int): The grid to round to. A positive/negative number
+            expands/shrinks the extent respectively.
 
     Returns:
         new_bounds (list): The new, rounded extent bounds.
     """
+    
+    if grid:
+        new_bounds = [
+            round_to(bounds[0], -grid), # xmin
+            round_to(bounds[1], -grid), # ymin
+            round_to(bounds[2], +grid), # xmax
+            round_to(bounds[3], +grid) # ymax
+        ]
 
-    new_bounds = [
-        round_to(bounds[0], -resolution), # xmin
-        round_to(bounds[1], -resolution), # ymin
-        round_to(bounds[2], +resolution), # xmax
-        round_to(bounds[3], +resolution) # ymax
-    ]
+        return new_bounds
+    
+    else: ### grid == 0
+        return bounds
 
-    return new_bounds
+def calc_transform(bounds, dst_resolution, src_crs, dst_crs, dst_round=0, tap=True):
+    """Calculates source and destination transforms from provided parameters.
+
+    Args:
+        bounds (list, tuple or polygon): A series of extent bounds (xmin, ymin,
+            xmax, ymax) or a polygon to calculate the extent from.
+        dst_resolution (int): Destination resolution in meters.
+        src_crs (str): Projection of the source.
+        dst_src (str): Projection of the destination.
+        dst_round (number, optional): The grid to round the extent to.
+        tap (bool, optional): Whether to align pixels to target resolution.
+
+    Returns:
+        dst_transform (dict): Destination transform.
+        dst_width (int): Width of the destination in pixels.
+        dst_height (int): Height of the destination in pixels.
+    """
+
+    if isinstance(bounds, (list, tuple)):
+        bounds = box(*bounds) ### Convert into polygon
+
+    if isinstance(bounds, Polygon):
+        if not src_crs:
+            src_crs = dst_crs
+
+        src_extent = gpd.GeoSeries(bounds, crs=src_crs)
+        dst_extent = round_extent(src_extent.to_crs(dst_crs).total_bounds, dst_round)
+
+    else:
+        raise TypeError("The bounding box needs to be list, tuple or polygon.")
+
+    ### Reprojection
+    dst_width = int((dst_extent[2]-dst_extent[0])/dst_resolution)
+    dst_height = int((dst_extent[3]-dst_extent[1])/dst_resolution)
+
+    dst_transform = rasterio.transform.from_bounds(
+        *dst_extent,
+        width=dst_width,
+        height=dst_height,
+    )
+
+    if tap:
+        return aligned_target(dst_transform, dst_width, dst_height, dst_resolution)
+    else:
+        return dst_transform, dst_width, dst_height
+
+def reproj_image(image, output, resolution, crs=None, round=0, tap=True, mode=0, verbose=False):
+    """Reprojects an image.
+
+    Args:
+        image (path): Input image. Its projection information has to be correct.
+        output (path): Output image.
+        resolution (int): Resolution of output image in meters.
+        crs (str, optional): Projection of the output image. If not given, uses
+            input projection.
+        round (number, optional): The grid to round the extent to.
+        tap (bool, optional): Whether to align pixels to target resolution.
+        mode (int, optional): Mode used for checking the output paths. Based on
+            the `check_path` function.
+        verbose (bool, optional): Whether to print current process status.
+
+    Returns:
+        None
+    """
+
+    check_path(Path(output).parent, mode)
+
+    with rasterio.open(image) as src:
+        profile = src.profile
+
+        dst_transform, dst_width, dst_height = calc_transform(
+            src.bounds,
+            resolution,
+            src.crs,
+            crs or src.crs,
+            round,
+            tap,
+        )
+        
+        count = src.count
+
+        profile.update(
+            transform=dst_transform,
+            width=dst_width,
+            height=dst_height,
+            crs=crs
+        )
+
+        with rasterio.open(output, 'w', **profile) as dst:
+            if verbose:
+                print(f"Bands wirtten: 0/{count}", end='\r')
+
+            for i in range(count):
+                dst_array, dst_transform = reproject(
+                    source=src.read(i+1),
+                    src_transform=src.transform,
+                    src_crs=src.crs,
+                    src_nodata=profile['nodata'],
+
+                    destination=np.zeros((dst_height, dst_width), profile['dtype']),
+                    dst_transform=dst_transform,
+                    dst_crs=crs,
+                    dst_nodata=profile['nodata'],
+                    dst_resolution=(resolution, resolution),
+
+                    resampling=Resampling.nearest,
+                    num_threads=4,
+                    warp_mem_limit=0,
+                )
+
+                dst.write(dst_array, i+1)
+
+                if verbose:
+                    print(f"Bands wirtten: {i+1}/{count}", end='\r')
+                
+            if verbose:
+                print()